backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.4
pydantic-settings==2.3.4
httpx==0.27.0
apscheduler==3.10.4
lightgbm==4.3.0
scikit-learn==1.5.0
joblib==1.4.2
numpy==1.26.4
python-dotenv==1.0.1
